Wrap Component in common module; move ModelDisplayHints.
See discussion in #927. getBodyList() and similar instantiations are now on
Model instead of Component. This allows us to move Component back to the
Common module.

ModelDisplayHints must be moved to Common, since Component relies on it.
