+++ OpenSim/Common/RegisterTypes_osimCommon.cpp
@@ -45,6 +45,8 @@
 
 #include "Reporter.h"
 
+#include "ModelDisplayHints.h"
+
 #include <string>
 #include <iostream>
 #include <exception>
@@ -94,6 +96,8 @@ OSIMCOMMON_API void RegisterTypes_osimCommon()
     Object::registerType( TableReporterVector() );
     Object::registerType( ConsoleReporter() );
     Object::registerType( ConsoleReporterVec3() );
+
+    Object::registerType( ModelDisplayHints());
 
     // TODO: temporarily map old NaturalCubicSpline (which wasn't a
     // natural cubic spline) to renamed SimmSpline class. Later we
