+++ OpenSim/Simulation/Model/ModelVisualPreferences.h
@@ -29,7 +29,7 @@
 #include <OpenSim/Simulation/osimSimulationDLL.h>
 #include "SimTKcommon.h"
 #include <OpenSim/Common/Property.h>
-#include <OpenSim/Simulation/Model/ModelDisplayHints.h>
+#include <OpenSim/Common/ModelDisplayHints.h>
 
 namespace OpenSim {
