+++ OpenSim/Simulation/Wrap/WrapCylinder.cpp
@@ -33,7 +33,7 @@
 #include <OpenSim/Common/SimmMacros.h>
 #include <OpenSim/Common/Mtx.h>
 #include <sstream>
-#include "OpenSim/Simulation/Model/ModelDisplayHints.h"
+#include <OpenSim/Common/ModelDisplayHints.h>
 //=============================================================================
 // STATICS
 //=============================================================================
