+++ OpenSim/Simulation/osimSimulation.h
@@ -26,7 +26,6 @@
 #include "Model/AnalysisSet.h"
 #include "Model/Bhargava2004MuscleMetabolicsProbe.h"
 #include "Model/Model.h"
-#include "Model/ModelDisplayHints.h"
 #include "Model/ModelVisualizer.h"
 #include "Model/ForceSet.h"
 #include "Model/BodyScale.h"
